Rename Internal module to Api
