+++ src/php/Api/Domain/PhelFnNormalizer.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Phel\Internal\Domain;
+namespace Phel\Api\Domain;
 
-use Phel\Internal\Infrastructure\PhelFnLoaderInterface;
-use Phel\Internal\Transfer\NormalizedPhelFunction;
+use Phel\Api\Infrastructure\PhelFnLoaderInterface;
+use Phel\Api\Transfer\NormalizedPhelFunction;
 use Phel\Lang\Keyword;
 
 final class PhelFnNormalizer implements PhelFnNormalizerInterface
